Add quickfix for null comparison against UnityObject
