+++ resharper/resharper-unity/test/data/CSharp/Daemon/Stages/PerformanceCriticalCodeAnalysis/InefficientCameraMainUsageWarningTest.cs
